# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Amadeus Node HTTP API
  description: |
    REST API for the Amadeus blockchain node, providing access to blockchain data,
    peer information, wallet operations, and smart contract interactions.

    This API enables developers to:
    - Query blockchain state and transactions
    - Manage wallet balances and transfers
    - Interact with smart contracts
    - Monitor network peers and validators

    ## Authentication
    Most endpoints are publicly accessible for reading blockchain data.
    Transaction submission requires proper cryptographic signatures.

    ## Base URL
    The API is typically available at `http://localhost` (port 80 by default, configurable via HTTP_PORT env var) when running a local node.
  version: 1.1.9
  contact:
    name: Amadeus Project
    url: https://github.com/amadeus-robot/node
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://72.9.144.110
    description: Amadeus seed node (primary)
  - url: http://167.235.169.185
    description: Amadeus seed node (secondary)
  - url: http://37.27.238.30
    description: Amadeus seed node (tertiary)
  - url: http://localhost
    description: Local development server (HTTP_PORT=80)
...
paths:
...
  /api/chain/height/{height}:
    get:
      tags: [chain]
      summary: Get entries by height
      description: Retrieves all blockchain entries at a specific height
      operationId: getEntriesByHeight
      parameters:
        - name: height
          in: path
          required: true
          description: Blockchain height to query
          schema:
            type: integer
            minimum: 0
            example: 12345
      responses:
        '200':
          description: Entries retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntriesResponse'
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X GET "http://72.9.144.110/api/chain/height/12345"
  /api/chain/height_with_txs/{height}:
    get:
      tags: [chain]
      summary: Get entries by height with transactions
      description: Retrieves blockchain entries at a specific height including all transactions
      operationId: getEntriesByHeightWithTxs
      parameters:
        - name: height
          in: path
          required: true
          description: Blockchain height to query
          schema:
            type: integer
            minimum: 0
            example: 12345
      responses:
        '200':
          description: Entries with transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntriesWithTxsResponse'
  /api/chain/tx/{txId}:
    get:
      tags: [transaction]
      summary: Get transaction by ID
      description: Retrieves a specific transaction by its ID
      operationId: getTransactionById
      parameters:
        - name: txId
          in: path
          required: true
          description: Transaction ID (hash)
          schema:
            type: string
            example: "abc123def456..."
      responses:
        '200':
          description: Transaction retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-codeSamples:
        - lang: curl
          label: cURL
          source: |
            curl -X GET "http://72.9.144.110/api/chain/tx/abc123def456..."
  /api/chain/tx_events_by_account/{account}:
    get:
      tags: [transaction]
      summary: Get transaction events by account
      description: |
        Retrieves transaction events for a specific account with optional filtering.
        Supports pagination and various filter options.
      operationId: getTransactionEventsByAccount
      parameters:
        - name: account
          in: path
          required: true
          description: Account public key (Base58 encoded)
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of results to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: Sort order for results
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: cursor_b58
          in: query
          description: Base58-encoded cursor for pagination
          schema:
            type: string
        - name: contract_b58
          in: query
          description: Filter by contract address (Base58 encoded)
          schema:
            type: string
        - name: function
          in: query
          description: Filter by contract function name
          schema:
            type: string
        - name: type
          in: query
          description: Filter by transaction type
          schema:
            type: string
            enum: [sent, recv]
      responses:
        '200':
          description: Transaction events retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionEventsResponse'
  /api/chain/txs_in_entry/{entryHash}:
    get:
      tags: [transaction]
      summary: Get transactions in entry
      description: Retrieves all transactions contained in a specific blockchain entry
      operationId: getTransactionsInEntry
      parameters:
        - name: entryHash
          in: path
          required: true
          description: Entry hash
          schema:
            type: string
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsInEntryResponse'
...
  /api/contract/validate_bytecode:
    post:
      tags: [contract]
      summary: Validate contract bytecode
      description: Validates WebAssembly contract bytecode for deployment
      operationId: validateContractBytecode
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: WebAssembly bytecode
      responses:
        '200':
          description: Bytecode validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BytecodeValidationResponse'
  /api/contract/get/{contractAddress}/{key}:
    get:
      tags: [contract]
      summary: Get contract state value
      description: Retrieves a specific value from contract storage
      operationId: getContractState
      parameters:
        - name: contractAddress
          in: path
          required: true
          description: Base58-encoded contract address
          schema:
            type: string
        - name: key
          in: path
          required: true
          description: Storage key to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Contract state retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Contract state value (format varies)
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum: [not_found, invalid_format, network_error]
          description: Error type
      required: [error]
...
    EntriesResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok]
        entries:
          type: array
          items:
            $ref: '#/components/schemas/BlockEntry'
      required: [error, entries]
    EntriesWithTxsResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok]
        entries:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/BlockEntry'
              - type: object
                properties:
                  txs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
      required: [error, entries]
    BlockEntry:
      type: object
      properties:
        hash:
          type: string
          description: Entry hash
        height:
          type: integer
          description: Block height
        timestamp:
          type: integer
          description: Block timestamp
        previous_hash:
          type: string
          description: Previous block hash
        merkle_root:
          type: string
          description: Merkle tree root
        signature:
          type: string
          description: Block signature
        mask:
          type: string
          description: Consensus mask
    TransactionResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok, not_found]
        transaction:
          $ref: '#/components/schemas/Transaction'
    Transaction:
      type: object
      properties:
        hash:
          type: string
          description: Transaction hash
        from:
          type: string
          description: Sender public key
        to:
          type: string
          description: Recipient address
        amount:
          type: string
          description: Transaction amount
        symbol:
          type: string
          description: Token symbol
        fee:
          type: string
          description: Transaction fee
        nonce:
          type: integer
          description: Transaction nonce
        timestamp:
          type: integer
          description: Transaction timestamp
        signature:
          type: string
          description: Transaction signature
        type:
          type: string
          enum: [transfer, contract_call, contract_deploy]
          description: Transaction type
    TransactionEventsResponse:
      type: object
      properties:
        cursor:
          type: string
          description: Pagination cursor for next page
        txs:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    TransactionsInEntryResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok]
        txs:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
      required: [error, txs]
...
    BytecodeValidationResponse:
      type: object
      properties:
        error:
          type: string
          enum: [ok, invalid_bytecode, too_large, unsupported_features]
        valid:
          type: boolean
          description: Whether bytecode is valid
        gas_estimate:
          type: integer
          description: Estimated gas cost for deployment
        warnings:
          type: array
          items:
            type: string
          description: Validation warnings
...