# http/src/routes/api/chain.rs
use crate::models::*;
use ama_core::Context;
use axum::{
    extract::{Path, State},
    Json,
};
use std::sync::Arc;
use utoipa;

// Router function not needed - routes defined in mod.rs

#[utoipa::path(
    get,
    path = "/api/chain/stats",
    summary = "Get blockchain statistics",
    description = "Retrieves current blockchain statistics including height, transaction counts, and network metrics.",
    responses(
        (status = 200, description = "Chain statistics retrieved successfully", body = ChainStatsResponse)
    ),
    tag = "chain"
)]
pub async fn get_chain_stats(State(ctx): State<Arc<Context>>) -> Json<ChainStatsResponse> {
    let stats = ChainStats {
        height: ctx.get_block_height(),
        total_transactions: 0, // placeholder - would need to query from fabric
        total_accounts: 0,     // placeholder - would need to query from fabric
        network_hash_rate: "0 H/s".to_string(), // placeholder
        difficulty: "0x0".to_string(),           // placeholder
    };

    Json(ChainStatsResponse::ok(stats))
}

#[utoipa::path(
    get,
    path = "/api/chain/tip",
    summary = "Get chain tip",
    description = "Retrieves the current tip (latest block) of the blockchain",
    responses(
        (status = 200, description = "Chain tip retrieved successfully", body = ChainTipResponse)
    ),
    tag = "chain"
)]
pub async fn get_chain_tip(State(_ctx): State<Arc<Context>>) -> Json<ChainTipResponse> {
    // placeholder implementation - would query current tip from consensus layer
    let entry = BlockEntry {
        hash: "0x1234567890abcdef".to_string(),
        height: 12345,
        timestamp: 1695123456,
        previous_hash: "0x0987654321fedcba".to_string(),
        merkle_root: "0xabcdef1234567890".to_string(),
        signature: "signature_data".to_string(),
        mask: "consensus_mask".to_string(),
    };

    Json(ChainTipResponse::ok(entry))
}

#[utoipa::path(
    get,
    path = "/api/chain/height/{height}",
    summary = "Get entries by height",
    description = "Retrieves blockchain entries at a specific height",
    responses(
        (status = 200, description = "Entries retrieved successfully", body = EntriesResponse)
    ),
    params(
        ("height" = u64, Path, description = "Block height to query", example = 12345)
    ),
    tag = "chain"
)]
pub async fn get_entries_by_height(
    State(_ctx): State<Arc<Context>>,
    Path(height): Path<u64>,
) -> Json<EntriesResponse> {
    // placeholder implementation - would query entries from fabric by height
    let entries = vec![BlockEntry {
        hash: format!("0x{:x}", height),
        height,
        timestamp: 1695123456 + height,
        previous_hash: format!("0x{:x}", height - 1),
        merkle_root: format!("0x{:x}root", height),
        signature: format!("sig_{}", height),
        mask: format!("mask_{}", height),
    }];

    Json(EntriesResponse::ok(entries))
}

#[utoipa::path(
    get,
    path = "/api/chain/height_with_txs/{height}",
    summary = "Get entries by height with transactions",
    description = "Retrieves blockchain entries at a specific height including all transactions",
    responses(
        (status = 200, description = "Entries with transactions retrieved successfully", body = EntriesWithTxsResponse)
    ),
    params(
        ("height" = u64, Path, description = "Block height to query", example = 12345)
    ),
    tag = "chain"
)]
pub async fn get_entries_by_height_with_txs(
    State(_ctx): State<Arc<Context>>,
    Path(height): Path<u64>,
) -> Json<EntriesWithTxsResponse> {
    // placeholder implementation
    let entry_with_txs = BlockEntryWithTxs {
        entry: BlockEntry {
            hash: format!("0x{:x}", height),
            height,
            timestamp: 1695123456 + height,
            previous_hash: format!("0x{:x}", height - 1),
            merkle_root: format!("0x{:x}root", height),
            signature: format!("sig_{}", height),
            mask: format!("mask_{}", height),
        },
        txs: vec![
            Transaction {
                hash: format!("tx_{}", height),
                from: "7EVUJfpnEqK32KrzUAaR4Yf26KgT66AWwW63xSHw5mbgdQhV8iwL1NkHMAqTi5Hv3h".to_string(),
                to: "6K9RwSR3gKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKg".to_string(),
                amount: "100.0".to_string(),
                symbol: "AMA".to_string(),
                fee: "0.1".to_string(),
                nonce: height,
                timestamp: 1695123456 + height,
                signature: format!("tx_sig_{}", height),
                tx_type: "transfer".to_string(),
            }
        ],
    };

    Json(EntriesWithTxsResponse::ok(vec![entry_with_txs]))
}

#[utoipa::path(
    get,
    path = "/api/chain/tx/{tx_id}",
    summary = "Get transaction by ID",
    description = "Retrieves a specific transaction by its ID",
    responses(
        (status = 200, description = "Transaction retrieved successfully", body = TransactionResponse),
        (status = 404, description = "Transaction not found", body = TransactionResponse)
    ),
    params(
        ("tx_id" = String, Path, description = "Transaction ID", example = "tx_12345")
    ),
    tag = "chain"
)]
pub async fn get_transaction_by_id(
    State(_ctx): State<Arc<Context>>,
    Path(tx_id): Path<String>,
) -> Json<TransactionResponse> {
    // placeholder implementation - would query transaction from fabric
    if tx_id.starts_with("tx_") {
        let transaction = Transaction {
            hash: tx_id.clone(),
            from: "7EVUJfpnEqK32KrzUAaR4Yf26KgT66AWwW63xSHw5mbgdQhV8iwL1NkHMAqTi5Hv3h".to_string(),
            to: "6K9RwSR3gKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKg".to_string(),
            amount: "100.0".to_string(),
            symbol: "AMA".to_string(),
            fee: "0.1".to_string(),
            nonce: 42,
            timestamp: 1695123456,
            signature: format!("{}_signature", tx_id),
            tx_type: "transfer".to_string(),
        };

        Json(TransactionResponse::ok(Some(transaction)))
    } else {
        Json(TransactionResponse::error("not_found"))
    }
}

#[utoipa::path(
    get,
    path = "/api/chain/tx_events_by_account/{account}",
    summary = "Get transaction events by account",
    description = "Retrieves transaction events for a specific account",
    responses(
        (status = 200, description = "Transaction events retrieved successfully", body = TransactionEventsResponse)
    ),
    params(
        ("account" = String, Path, description = "Account public key", example = "7EVUJfpnEqK32KrzUAaR4Yf26KgT66AWwW63xSHw5mbgdQhV8iwL1NkHMAqTi5Hv3h")
    ),
    tag = "chain"
)]
pub async fn get_transaction_events_by_account(
    State(_ctx): State<Arc<Context>>,
    Path(account): Path<String>,
) -> Json<TransactionEventsResponse> {
    // placeholder implementation - would query events from fabric
    let transactions = vec![
        Transaction {
            hash: format!("event_tx_1_{}", account.chars().take(8).collect::<String>()),
            from: account.clone(),
            to: "6K9RwSR3gKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKg".to_string(),
            amount: "50.0".to_string(),
            symbol: "AMA".to_string(),
            fee: "0.1".to_string(),
            nonce: 1,
            timestamp: 1695123456,
            signature: "event_signature_1".to_string(),
            tx_type: "transfer".to_string(),
        },
        Transaction {
            hash: format!("event_tx_2_{}", account.chars().take(8).collect::<String>()),
            from: "7EVUJfpnEqK32KrzUAaR4Yf26KgT66AWwW63xSHw5mbgdQhV8iwL1NkHMAqTi5Hv3h".to_string(),
            to: account,
            amount: "25.0".to_string(),
            symbol: "AMA".to_string(),
            fee: "0.1".to_string(),
            nonce: 2,
            timestamp: 1695123466,
            signature: "event_signature_2".to_string(),
            tx_type: "transfer".to_string(),
        },
    ];

    Json(TransactionEventsResponse {
        cursor: Some("next_cursor_placeholder".to_string()),
        txs: transactions,
    })
}

#[utoipa::path(
    get,
    path = "/api/chain/txs_in_entry/{entry_hash}",
    summary = "Get transactions in entry",
    description = "Retrieves all transactions within a specific blockchain entry",
    responses(
        (status = 200, description = "Transactions retrieved successfully", body = TransactionsInEntryResponse)
    ),
    params(
        ("entry_hash" = String, Path, description = "Entry hash", example = "0x1234567890abcdef")
    ),
    tag = "chain"
)]
pub async fn get_transactions_in_entry(
    State(_ctx): State<Arc<Context>>,
    Path(entry_hash): Path<String>,
) -> Json<TransactionsInEntryResponse> {
    // placeholder implementation - would query transactions from specific entry
    let transactions = vec![
        Transaction {
            hash: format!("entry_tx_1_{}", entry_hash.chars().take(8).collect::<String>()),
            from: "7EVUJfpnEqK32KrzUAaR4Yf26KgT66AWwW63xSHw5mbgdQhV8iwL1NkHMAqTi5Hv3h".to_string(),
            to: "6K9RwSR3gKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKgKg".to_string(),
            amount: "100.0".to_string(),
            symbol: "AMA".to_string(),
            fee: "0.1".to_string(),
            nonce: 42,
            timestamp: 1695123456,
            signature: "entry_tx_signature_1".to_string(),
            tx_type: "transfer".to_string(),
        }
    ];

    Json(TransactionsInEntryResponse::ok(transactions))
}